% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_generisk.R
\name{bootstrap_generisk}
\alias{bootstrap_generisk}
\title{bootstrap a generisk model}
\usage{
bootstrap_generisk(generisk_obj, B, start = "fit")
}
\arguments{
\item{generisk_obj}{object from generisk}

\item{B}{number of bootstrap}

\item{start}{init values for parameters, if "fit" starts from the initial fit, otherwise starts from population incidence}
}
\value{
to be completed
}
\description{
bootstrap a generisk model
}
\examples{
#to be completed
}
