% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpedigrees.R
\name{checkpedigrees}
\alias{checkpedigrees}
\title{function largely inspired from the "familycheck" fct of the Kinship package}
\usage{
checkpedigrees(famid, id, father.id, mother.id)
}
\arguments{
\item{famid}{family identification number (vector)}

\item{id}{individual identification number (vector)}

\item{father.id}{father id}

\item{mother.id}{mother id}
}
\value{
to be completed
}
\description{
function largely inspired from the "familycheck" fct of the Kinship package
}
\examples{
#to be completed

}
