% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{function to compare two models}
\usage{
compareModels(model1, model2)
}
\arguments{
\item{model1}{a model object returned by the generisk() function}

\item{model2}{a model object returned by the generisk() function}
}
\value{
a list of two: likelihood ratio and p-value
}
\description{
function to compare two models
}
\examples{
#to be completed

}
