% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_loglik.R
\name{compute_loglik_boot}
\alias{compute_loglik_boot}
\title{compute the log-likelihood from a bootstrapped model}
\usage{
compute_loglik_boot(generisk_obj)
}
\arguments{
\item{generisk_obj}{an object returned by the generisk function}
}
\value{
to be completed
}
\description{
compute the log-likelihood from a bootstrapped model
}
\examples{
#to be completed
}
