% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generisk.R
\name{generisk}
\alias{generisk}
\title{estimate age-risk (penetrance) functions from family data}
\usage{
generisk(
  Ft.pop,
  LIK.method = "GRL",
  DATA,
  FIT.pars = list(cancer = list(penet.model = "np", agenodes = c(30, 50, 70), implic.loci
    = c(MMR = TRUE), inheritance = c(MMR = "dominant"), gender.effect = c(MMR = TRUE))),
  fA = c(MMR = 0.001),
  B = 0,
  rel.tol = 1e-06,
  rel.tol.boot = 1e-06,
  approxFt.aa = TRUE,
  approx.np = "spline",
  first.cancer = FALSE,
  init.weibull = NULL
)
}
\arguments{
\item{Ft.pop}{a list of 121*2 matrices of diseases incidences (length = ndis)}

\item{LIK.method}{if GRL : first element of fA must be the allele frequency for the gene under selection}

\item{DATA}{a N*(6 + nloci + 2*ndis) matrix}

\item{FIT.pars}{a list of diseases parameters (length = ndis) (see details)}

\item{fA}{a vector of pop. allele frequencies of tested loci (order must corresponds to that of DATA columns)}

\item{B}{option to calculate 95\% Bootstrap estimates # number of bootstrap resample (B=0 means no bootstrap)}

\item{rel.tol}{precision for the likelihood}

\item{rel.tol.boot}{precision for the likelihood in the bootstrap}

\item{approxFt.aa}{option to calculate Ft.aa}

\item{approx.np}{linear or spline}

\item{first.cancer}{keep only the first cancer for individuals with multiple cancers ?}

\item{init.weibull}{initial values for Weibull model}
}
\value{
to be completed
}
\description{
generisk() is the main function of the generisk package.
It performs penetrance estimation using maximum likelihood with the GRL (Genotype Restricted Likelihood),
PEL (Proband Excluded Likelihood), and PL (Prospective Likelihood) methods.
It relies on the Elston-Stewart algorithm to factorizing family data.
Different options are implemented for parametrization.
}
\examples{
#to be completed

}
