% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generisk.R
\name{plot_generisk}
\alias{plot_generisk}
\title{plot penetrance curves estimated with the generisk() function}
\usage{
plot_generisk(
  x,
  type = "absolute",
  conf = 0.95,
  add = FALSE,
  multiple = FALSE,
  ymax = 1,
  xmax = 100,
  allcurves = FALSE,
  legendloc = "right",
  legendtext = NULL,
  cols = rep(c("#87CEFA", "#FF0000", "#00CD00", "#00CED1", "#E9967A", "#006400",
    "#FF8C00", "#00008B", "#8B008B", "#008B8B"), 3),
  main = "",
  show.other = FALSE
)
}
\arguments{
\item{x}{an object returned by the generisk function}

\item{type}{absolute or relative}

\item{conf}{show bootstrapped 95\% confidence intervals}

\item{add}{add to existing graph}

\item{multiple}{multiple graphs in case of multiple phenotypes}

\item{ymax}{maximum value for y axis}

\item{xmax}{maximum value for x axis}

\item{allcurves}{show all bootstrapped penetrance curves}

\item{legendloc}{location of the legend}

\item{legendtext}{text of the legend}

\item{cols}{colors}

\item{main}{main title of the graph}

\item{show.other}{show the group "other"}
}
\value{
to be completed
}
\description{
plot penetrance curves estimated with the generisk() function
}
\examples{
#to be completed

}
