% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statdesc.R
\name{statdesc_table}
\alias{statdesc_table}
\title{summary statistics table of the number of at-risk and the number of cancers by age and genotype from a generisk object}
\usage{
statdesc_table(x)
}
\arguments{
\item{x}{an object returned by the generisk function}
}
\value{
to be completed
}
\description{
summary statistics table of the number of at-risk and the number of cancers by age and genotype from a generisk object
}
\examples{
#to be completed

}
