% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpedigrees.R
\name{checkpedigrees}
\alias{checkpedigrees}
\title{function checking the consistency of family data with help of the pedtools::ped() function}
\usage{
checkpedigrees(fams)
}
\arguments{
\item{fams}{a dataframe with columns famid, id, fatherid, motherid, sex}
}
\value{
to be completed
}
\description{
function checking the consistency of family data with help of the pedtools::ped() function
}
\examples{
#to be completed

}
