% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cumrisk.R
\name{compute_cumrisk}
\alias{compute_cumrisk}
\title{compute cumulative risk over multiple traits from a bootstrapped generisk call object}
\usage{
compute_cumrisk(
  x,
  conf = 0.95,
  ages = seq(30, 70, 10),
  traits = names(x$params$FIT.pars),
  digit_abs = 1,
  digit_rr = 1,
  digit_abs_agec = 1
)
}
\arguments{
\item{x}{an object returned by the generisk function}

\item{conf}{show bootstrapped 95\% confidence intervals}

\item{ages}{at which ages ?}

\item{traits}{over which traits ?}

\item{digit_abs}{digit for absolute risks}

\item{digit_rr}{digit for relative risks}

\item{digit_abs_agec}{digit for absolute risks by agec}
}
\value{
to be completed
}
\description{
compute cumulative risk over multiple traits from a bootstrapped generisk call object
}
\examples{
#to be completed

}
