% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generisk.R
\name{generisk}
\alias{generisk}
\title{generisk is the main function of the generisk package.}
\usage{
generisk(
  Ft.pop,
  LIK.method = "GRL",
  DATA,
  FIT.pars = list(cancer = list(penet.model = "np", agenodes = c(30, 50, 70), implic.loci
    = c(MMR = TRUE), inheritance = c(MMR = "dominant"), gender.effect = c(MMR = TRUE))),
  fA = c(MMR = 0.001),
  B = 0,
  rel.tol = 1e-06,
  approxFt.aa = TRUE,
  n.points = 5,
  approx.np = "spline",
  ncores = 4,
  imput_ageNA = FALSE
)
}
\arguments{
\item{Ft.pop}{a list of 121*2 matrices of diseases incidences (length = ndis)}

\item{LIK.method}{if GRL : first element of fA must be the allele frequency for the gene under selection}

\item{DATA}{a N*(6 + nloci + 2*ndis) matrix}

\item{FIT.pars}{a list of diseases parameters (length = ndis) (see details)}

\item{fA}{a vector of pop. allele frequencies of tested loci (order must corresponds to that of DATA columns)}

\item{B}{option to calculate 95\% Bootstrap estimates # number of bootstrap resample (B=0 means no bootstrap)}

\item{rel.tol}{precision for the likelihood}

\item{approxFt.aa}{option to calculate Ft.aa}

\item{n.points}{number of points used for the non parametric penetrance function}

\item{approx.np}{linear or spline}

\item{ncores}{number of cores used for parallel computing}

\item{imput_ageNA}{logical: imputation of missing ages at last news ?}
}
\value{
to be completed
}
\description{
generisk is the main function of the generisk package.
}
\examples{
#to be completed

}
