% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generisk.R
\name{generisk}
\alias{generisk}
\title{estimate age-risk (penetrance) functions from family data}
\usage{
generisk(
  Ft.pop,
  LIK.method = "GRL",
  DATA,
  FIT.pars = list(cancer = list(penet.model = "np", agenodes = c(30, 50, 70), implic.loci
    = c(MMR = TRUE), inheritance = c(MMR = "dominant"), gender.effect = c(MMR = TRUE))),
  fA = c(MMR = 0.001),
  B = 0,
  rel.tol = 1e-06,
  rel.tol.boot = 1e-06,
  approxFt.aa = TRUE,
  approx.np = "spline",
  multi.pheno = "all",
  init.weibull = NULL,
  ncores = 1,
  imput_missing_age_last_news = FALSE
)
}
\arguments{
\item{Ft.pop}{a list of 121*2 matrices of diseases incidences (length = ndis)}

\item{LIK.method}{if GRL : first element of fA must be the allele frequency for the gene under selection}

\item{DATA}{a N*(6 + nloci + 2*ndis) matrix}

\item{FIT.pars}{a list of diseases parameters (length = ndis) (see details)}

\item{fA}{a vector of pop. allele frequencies of tested loci (order must corresponds to that of DATA columns)}

\item{B}{option to calculate 95\% Bootstrap estimates # number of bootstrap resample (B=0 means no bootstrap)}

\item{rel.tol}{precision for the likelihood}

\item{rel.tol.boot}{precision for the likelihood in the bootstrap}

\item{approxFt.aa}{option to calculate Ft.aa}

\item{approx.np}{linear or spline}

\item{multi.pheno}{strategy in case of multiple diseases (can be "first" or "all")}

\item{init.weibull}{initial values for Weibull model}

\item{ncores}{number of cores used for parallel computing}

\item{imput_missing_age_last_news}{logical: imputation of missing ages at last news ?}
}
\value{
to be completed
}
\description{
generisk() is the main function of the generisk package.
It performs penetrance estimation using maximum likelihood with the GRL method.
It relies on the Elston-Stewart algorithm to factorizing family data.
Different options are implemented for parametrization.
}
\examples{
#to be completed

}
