% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nloglik.r
\name{nloglik}
\alias{nloglik}
\title{Title}
\usage{
nloglik(x, return_lklbyfam = FALSE, ...)
}
\arguments{
\item{x}{parameters of the penetrance curve}

\item{return_lklbyfam}{return a table with LKL by family}

\item{...}{internal args}
}
\value{
to be completed
}
\description{
Title
}
\examples{
#to be completed

}
