% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_generisk.R
\name{summarize_generisk}
\alias{summarize_generisk}
\title{summarize object estimated with the generisk() function}
\usage{
summarize_generisk(
  x,
  conf = 0.95,
  ages = seq(10, 70, 10),
  digit_abs = 1,
  digit_rr = 1,
  digit_abs_agec = 1
)
}
\arguments{
\item{x}{an object returned by the generisk function}

\item{conf}{show bootstrapped 95\% confidence intervals}

\item{ages}{at which ages to summarize ?}

\item{digit_abs}{digit for absolute risks}

\item{digit_rr}{digit for relative risks}

\item{digit_abs_agec}{digit for absolute risks by agec}
}
\value{
to be completed
}
\description{
summarize object estimated with the generisk() function
}
\examples{
#to be completed

}
